refactor: viewmodel에서 bpm 데이터를 넘겨주도록 수정한다
