+++ client/src/components/blocks/Bpm/index.tsx
@@ -1,45 +1,23 @@
-import React, { FormEvent, useEffect, useRef, useState } from "react";
+import React, { FormEventHandler } from "react";
 
 import Span from "components/atoms/Span";
-import { getSliderHandlePosition } from "utils/slider";
 import "./style.scss";
 
-// interface BpmProps {
-//   inputHandler: FormEventHandler;
-// }
+interface BpmProps {
+  bpm: number;
+  inputHandler: FormEventHandler;
+  sliderRef: React.RefObject<HTMLInputElement>;
+  sliderValueRef: React.RefObject<HTMLSpanElement>;
+}
 
-function Bpm() {
-  const [sliderValue, setSliderValue] = useState(60);
-  const sliderRef = useRef<HTMLInputElement>(null);
-  const sliderValueRef = useRef<HTMLSpanElement>(null);
-
-  const inputHandler = (e: FormEvent) => {
-    setSliderValue(Number((e.target as HTMLInputElement).value));
-
-    (sliderValueRef.current as HTMLSpanElement).style.left = getSliderHandlePosition(
-      0,
-      240,
-      Number((e.target as HTMLInputElement).value),
-      Number(sliderRef.current?.getBoundingClientRect().width) - 40,
-    );
-  };
-
-  useEffect(() => {
-    (sliderValueRef.current as HTMLSpanElement).style.left = getSliderHandlePosition(
-      0,
-      240,
-      Number(sliderRef.current!.value),
-      Number(sliderRef.current?.getBoundingClientRect().width) - 40,
-    );
-  }, []);
-
+function Bpm({ bpm, inputHandler, sliderRef, sliderValueRef }: BpmProps) {
   return (
     <div id="bpm">
       <Span>BPM</Span>
       <div className="slide-container">
         <input className="slider" type="range" min="0" max="240" step="1" onInput={inputHandler} ref={sliderRef} />
         <span id="value" ref={sliderValueRef}>
-          {sliderValue}
+          {bpm}
         </span>
         <div className="slider-bottom">
           <span>30</span>
